+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+## [0.8.0] - 2024-11-11
+
 ### Fixed
 
 -   Fixed the env marker evaluation
@@ -100,7 +102,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/spapanik/phosphorus/compare/v0.7.0...main
+[Unreleased]: https://github.com/spapanik/phosphorus/compare/v0.8.0...main
+[0.8.0]: https://github.com/spapanik/phosphorus/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/spapanik/phosphorus/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/spapanik/phosphorus/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/spapanik/phosphorus/compare/v0.4.1...v0.5.0